Refactor of run_devnet_in_background to pass linter
Extract run_devnet_in_background to new conftest.py file
